[Maintenance] [th01] Replace generic C types with more specific ones
In all places visited during the next 6 pushes: The resident structure
and copies of its values, the packfile implementation, boss entities,
rendering font ROM glyphs to VRAM, and overall inconsistent code
between the three binaries.

Part of P0229, funded by Ember2528.

+++ th01/core/resstuff.cpp
@@ -44,7 +44,7 @@ int resident_stuff_get(
 	int8_t& bombs,
 	int8_t& start_lives_extra,
 	unsigned long& rand,
-	long& continues_total,
+	int32_t& continues_total,
 	int& stage_id
 )
 {
